feed cache

+++ download.php
@@ -32,9 +32,11 @@ $converter->convert();
 
 $videos = glob(__DIR__ . '/static/*.mp4');
 foreach ($videos as $video) {
-    $age = time() . filemtime($video);
-    if ($age > 259200) {
-        unlink($video);
+    if (file_exists($video)) {
+        $age = time() . filemtime($video);
+        if ($age > 259200 && file_exists($video)) {
+            unlink($video);
+        }
     }
 }
